Use floats for vertex values

+++ src/render/forward/forward.vert
@@ -1,7 +1,7 @@
 #version 450
 
-layout(location = 0) in ivec4 a_Pos;
-layout(location = 1) in ivec4 a_Normal;
+layout(location = 0) in vec4 a_Pos;
+layout(location = 1) in vec4 a_Normal;
 
 layout(location = 0) out vec3 v_Normal;
 layout(location = 1) out vec4 v_Position;
